VectorSearchService/pyproject.toml
[tool.poetry]
name = "vector-search"
version = "0.1.0"
description = "FastAPI vector search with pgvector and OpenAI embeddings"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
openai = "^1.30.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}
pgvector = "^0.2.0"
pydantic = "^2.6.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
ruff = "^0.3.0"
pytest = "^8.0.0"
mypy = "^1.9.0"
pre-commit = "^3.6.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88 